extract : extract src path
