+++ frontend/src/components/home/RoomForm.vue
@@ -40,7 +40,6 @@ const onOk = () => {
   (formRef as any)?.value
     .validateFields()
     .then((values: any) => {
-      // TODO: 表单不为空 判断
       request.post('/api/room/add', toRaw(formState)).then((res: any) => {
         emit('addRoom', toRaw(formState));
         message.success(res.data.message)
